+++ eSlayerHelpers/src/erand.cpp
@@ -12,7 +12,7 @@ int eRand::rand() {
     return sDist(sRng);
 }
 
-double eRand::randF(const double min, const double max) {
+float eRand::randF(const float min, const float max) {
     const int margin = 100000;
     return min + (max - min)*(rand() % (margin + 1))/margin;
 }
