+++ eSlayerServer/src/eattackaction.cpp
@@ -19,14 +19,14 @@ eAttackAction::sCreate(
     } else {
         return nullptr;
     }
-    const double iasItem = 0.; // all items speed modifiers
-    const double eias = std::floor(iasItem*120./(iasItem + 120.));
-    const double sias = 0.; // skill increased attack speed
-    const double wsm = 0.; // weapon speed modifier
+    const float iasItem = 0.f; // all items speed modifiers
+    const float eias = std::floor(iasItem*120.f/(iasItem + 120.f));
+    const float sias = 0.f; // skill increased attack speed
+    const float wsm = 0.f; // weapon speed modifier
     const int animLen = data.animFrames(anim);
-    const double animSpeed = 256.;
-    const double animRate = 100.;
-    const int frames = int(std::ceil(256.*animLen/std::floor(animSpeed*(animRate + sias + eias - wsm)/100.))) - 1;
+    const float animSpeed = 256.f;
+    const float animRate = 100.f;
+    const int frames = int(std::ceil(256.f*animLen/std::floor(animSpeed*(animRate + sias + eias - wsm)/100.f))) - 1;
     const auto result = std::make_shared<eAttackAction>(
         unit, area);
     result->setup(anim, frames, a);
