+++ eSlayerServer/src/eclientaction.h
@@ -7,7 +7,7 @@ class eClientAction : public eComplexAction {
 public:
     using eComplexAction::eComplexAction;
 
-    void increment(const double by) override;
+    void increment(const float by) override;
     void decide() override;
 
     void attack(const std::shared_ptr<eServerUnit>& target);
